backend/requirements.txt
aiofiles==0.4.0
alembic==0.9.9
astroid==1.6.4
atomicwrites==1.3.0
attrs==18.2.0
blinker==1.4
click==6.7
dnspython==1.16.0
grpcio==1.18.0
h11==0.8.1
h2==3.1.0
hpack==3.0.0
Hypercorn==0.5.2
hyperframe==5.2.0
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
monotonic==1.5
more-itertools==5.0.0
multidict==4.5.2
mypy==0.660
mypy-extensions==0.4.1
pluggy==0.8.1
prometheus-client==0.5.0
prometheus-flask-exporter==0.5.1
protobuf==3.6.1
pulsar-client==2.2.1
py==1.7.0
py-dateutil==2.2
pylint==1.9.1
PyMySQL==0.9.3
pytest==4.2.0
pytest-mock==1.10.1
python-dateutil==2.7.3
python-editor==1.0.3
pytoml==0.1.20
Quart==0.6.12
Quart-CORS==0.1.2
six==1.12.0
sortedcontainers==2.1.0
SQLAlchemy==1.2.7
typed-ast==1.2.0
typing-extensions==3.7.2
Werkzeug==0.14.1
wrapt==1.10.11
wsproto==0.13.0
